Finished Minimap Refactoring
Implemented a larger minimap size, adjusted the sonar effect to fit the player's map, refactored the minimap to be handled modularly through a manager. All of the objects relating to the minimap are prefabs now instead of being dependent on the scene.
STILL TODO: Implement map tracking functionality in order for the large map to still track the player.